Implement __getitem__ function for the ChangeDetectionDataset class
In this commit I implement the functionality of getitem function. I will create all the possible combinations of images from the same AOI for our dataset and in case of training, it will return a crop, while in case of validation or testing it will return the whole image. Also, if in training mode, it will visit the same combination self.no_of_crops_per_combination times.

There is no data augmentation for now, it will be added in future commits.